# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bac Thinh API docs",
    "description": "Mo ta API service",
    "license": {
      "name": "API License",
      "url": "http://nguyenkiem.vn/license"
    },
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Server test"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/masses/{id}": {
      "get": {
        "tags": ["Mass Management"],
        "summary": "Get mass by ID",
        "description": "Retrieve a specific mass",
        "operationId": "getMassById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mass found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MassResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Mass not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Mass Management"],
        "summary": "Update a mass",
        "description": "Update an existing mass",
        "operationId": "updateMass",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMassRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Mass updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MassResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Mass not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Mass Management"],
        "summary": "Delete a mass",
        "description": "Delete a mass",
        "operationId": "deleteMass",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Mass deleted successfully"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Mass not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/masses/daily-schedule/{dailyScheduleId}": {
      "get": {
        "tags": ["Mass Management"],
        "summary": "Get masses by daily schedule",
        "description": "Get all masses for a specific daily schedule",
        "operationId": "getMassesByDailyScheduleId",
        "parameters": [
          {
            "name": "dailyScheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Masses found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MassResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Mass Management"],
        "summary": "Create a new mass",
        "description": "Create a new mass for a specific daily schedule",
        "operationId": "createMass",
        "parameters": [
          {
            "name": "dailyScheduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMassRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Mass created successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MassResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Daily schedule not found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/masses/by-date": {
      "get": {
        "tags": ["Mass Management"],
        "summary": "Get masses by date",
        "description": "Get all masses for a specific date",
        "operationId": "getMassesByDate",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Masses found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MassResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masses/by-date-range": {
      "get": {
        "tags": ["Mass Management"],
        "summary": "Get masses by date range",
        "description": "Get all masses within a date range",
        "operationId": "getMassesByDateRange",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Masses found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MassResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/masses/by-celebrant": {
      "get": {
        "tags": ["Mass Management"],
        "summary": "Get masses by celebrant",
        "description": "Get all masses for a specific celebrant within a date range",
        "operationId": "getMassesByCelebrant",
        "parameters": [
          {
            "name": "celebrant",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Masses found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MassResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseVoid"
                    },
                    {
                      "$ref": "#/components/schemas/ApiResponseString"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateMassRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "time": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "type": {
            "type": "string"
          },
          "celebrant": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "specialName": {
            "type": "string"
          },
          "isSolemn": {
            "type": "boolean"
          }
        },
        "required": ["celebrant", "time", "type"]
      },
...
      "MassResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "time": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "type": {
            "type": "string"
          },
          "celebrant": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "specialName": {
            "type": "string"
          },
          "isSolemn": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateMassRequest": {
        "type": "object",
        "properties": {
          "time": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "type": {
            "type": "string"
          },
          "celebrant": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "specialName": {
            "type": "string"
          },
          "isSolemn": {
            "type": "boolean"
          }
        },
        "required": ["celebrant", "time", "type"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}